+++ src/tools/DeepAnalysis/getUnknownTextMap.ts
@@ -11,11 +11,12 @@ import type {
 } from '../../AhkSymbol/CAhkFunc';
 import type { TGlobalVal, TGValMap } from '../../core/ParserTools/ahkGlobalDef';
 import type { TTokenStream } from '../../globalEnum';
-import { hoverAVar } from '../Built-in/A_Variables.tools';
+import { AVariablesMDMap } from '../Built-in/A_Variables.tools';
 import { Bi_VarMDMap } from '../Built-in/BiVariables.tools';
 import { CommandMDMap } from '../Built-in/Command.tools';
 import { otherKeyword2MD } from '../Built-in/otherKeyword2.tools';
 import { StatementMDMap } from '../Built-in/statement.tools';
+import { WinTitleMDMap } from '../Built-in/WinTitle/WinTitleParameter.tools';
 import { newC502 } from './FnVar/def/c502';
 
 function pushRef(
@@ -128,12 +129,14 @@ export function getUnknownTextMap(
 
             if (
                 !textMap.has(wordUp) && (
-                    CommandMDMap.has(wordUp) || (hoverAVar(wordUp) !== undefined) || StatementMDMap.has(wordUp)
+                    CommandMDMap.has(wordUp)
+                    || AVariablesMDMap.has(wordUp) || StatementMDMap.has(wordUp)
                     || Bi_VarMDMap.has(wordUp)
                     || (/^_+$/u).test(wordUp) // str
                     || (/^\d+$/u).test(wordUp) // just number
                     || (/^0X[\dA-F]+$/u).test(wordUp) // NumHexConst = 0 x [0-9a-fA-F]+
                     || otherKeyword2MD.has(wordUp)
+                    || WinTitleMDMap.has(wordUp)
                     /*
                      * let decLiteral: number = 6;
                      * let hexLiteral: number = 0xf00d;